image-processing-suite/requirements.txt
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
tqdm>=4.64.0
colorama>=0.4.4
piexif>=1.1.3
wand>=0.6.10
scikit-image>=0.19.0
matplotlib>=3.5.0
python-magic>=0.4.24
